Refactored Tile class. Now a Square class for "terrain squares" Began work on the Board class. The board is a subclass of Grid, with Kingdomino rules that restrict placement.
The playable area of the board is a 5x5 grid, which can be built out in any direction. This means, starting from the center tile, there is a 9x9 grid of playable space. To conserve screen space, only a 7x7 grid will be displayed / used. When a player builds toward one side, their kingdom will be recentered.