Added a game visualization
Added Display_game_state() to the Play class, which displays game information to the players. Scattered print statements now update the self.messages field.

This included updating the __str__ method of Board, to include chess-style indexes that are easier for human use. Board now has a helper function _chess_indexed to convert those into coordinates.

Also fixed an error when trying to undo the first action in the game.